Rename client.py to pylandax.py
